requirements.txt
aiofiles==0.8.0
anyio==3.5.0
asgiref==3.5.0
attrs==21.4.0
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
cryptography==36.0.1
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.74.1
h11==0.13.0
httptools==0.3.0
idna==3.3
iniconfig==1.1.1
jose==1.0.0
mysql-connector-python==8.0.28
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.7
pytest==7.0.1
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==6.0
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
stripe==2.66.0
tomli==2.0.1
typing-extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
uvloop==0.16.0
watchgod==0.7
websockets==10.2
databases~=0.5.5
asyncmy~=0.2.3
